feat: replace Offer type to offer.types.ts
