+++ project/src/types/offer.types.ts
@@ -1,0 +1,34 @@
+export interface Offer {
+  bedrooms: number,
+  city: {
+    location: {
+      latitude: number,
+      longitude: number,
+      zoom: number,
+    },
+    name: string,
+  },
+  description: string,
+  goods: string[],
+  host: {
+    avatarUrl: string,
+    id: number,
+    isPro: boolean,
+    name: string,
+  },
+  id: number,
+  images: string[],
+  isFavorite: boolean,
+  isPremium: boolean,
+  location: {
+    latitude: number,
+    longitude: number,
+    zoom: number,
+  },
+  maxAdults: number,
+  previewImage: string,
+  price: number,
+  rating: number,
+  title: string,
+  type: string,
+}
